Move top-level interfaces to a separate namespace/dir
This will help with code organization
