+++ docs/reference/Interfaces/ScribeInterface.md
@@ -1,0 +1,13 @@
+# Libgossamer API Reference - ScribeInterface
+
+* **Fully Qualified Interface Name**: `ParagonIE\Gossamer\Interfaces\ScribeInterface`
+
+## Interface Methods
+
+### `publish()`
+
+**Arguments**:
+
+1. [`SignedMessage`](../Protocol/SignedMessage.md) 
+
+Returns a `bool`.
