+++ docs/reference/Interfaces/SourceInterface.md
@@ -1,6 +1,6 @@
 # Libgossamer API Reference - SourceInterface
 
-* **Fully Qualified Interface Name**: `ParagonIE\Gossamer\SourceInterface`
+* **Fully Qualified Interface Name**: `ParagonIE\Gossamer\Interfaces\SourceInterface`
 
 ## Interface Methods
 
@@ -10,4 +10,4 @@
 
 1. `string` - Hash
 
-**Returns** a [`Response`](Response.md).
+**Returns** a [`Response`](../Response.md).
