+++ docs/reference/README.md
@@ -11,6 +11,15 @@ writing code that needs to use the library.
   * [PDO](Db/PDO.md)
 * [Http](Http)
   * [Guzzle](Http/Guzzle.md)
+* [Interfaces](Interfaces)
+  * [DbInterface](Interfaces/DbInterface.md)
+  * [CryptoBakendInterface](Interfaces/CryptoBackendInterface.md)
+  * [HttpInterface](Interfaces/HttpInterface.md)
+  * [LedgerInterface](Interfaces/LedgerInterface.md)
+  * [LedgerVerifierInterface](Interfaces/LedgerVerifierInterface.md)
+  * [ScribeInterface](Interfaces/ScribeInterface.md)
+  * [SourceInterface](Interfaces/SourceInterface.md)
+  * [VerifierInterface](Interfaces/VerifierInterface.md)
 * [Protocol](Protocol)
   * [Action](Protocol/Action.md)
   * [Message](Protocol/Message.md)
@@ -26,13 +35,7 @@ writing code that needs to use the library.
   * [Chronicle](Source/Chronicle.md)
 * [Verifier](Verifier)
   * [Chronicle](Verifier/Chronicle.md)
-* [DbInterface](DbInterface.md)
-* [CryptoBakendInterface](CryptoBackendInterface.md)
 * [GossamerException](GossmerException.md)
-* [HttpInterface](HttpInterface.md)
-* [LedgerInterface](LedgerInterface.md)
 * [Response](Response.md)
-* [ScribeInterface](ScribeInterface.md)
 * [Synchronizer](Synchronizer.md)
 * [Util](Util.md)
-* [VerifierInterface](VerifierInterface.md)
