--- docs/reference/ScribeInterface.md
@@ -1,13 +1,0 @@
-# Libgossamer API Reference - ScribeInterface
-
-* **Fully Qualified Interface Name**: `ParagonIE\Gossamer\ScribeInterface`
-
-## Interface Methods
-
-### `publish()`
-
-**Arguments**:
-
-1. [`SignedMessage`](Protocol/SignedMessage.md) 
-
-Returns a `bool`.
